Use opener for platform specific launching code
the current implementation of jolly detects what platform we are on and
uses xdg-open or explorer to open the file as appropriate.

However, there are some corner cases on Windows platforms in terms of
how the explorer.exe handles shell arguments, which means that urls
with query strings were not being handled properly. The solution is to start
using a standalone crate `opener` to handle opening of the files. This works
for all 3 platforms and allows simplifying the url logic.