Modified windows subshell spawning to avoid creating console window (#5)
When running a command on windows, windows needs to decide whether or not to create a console window that appears.

Since Jolly is a GUI application that does not have a console, this
 was appearing whenever system entries were executed.

However, this was only happening for release builds, since debug build
still create a console to show log output. It was further confused by
the fact that jolly is usually developed from the command line, so the
fact that debug builds were popping up a console was also not noticed,
since launching Jolly from the command prompt would hide this extra
window.

This changes the logic so that windows are never spawned with a
console attached, in debug or release mode.