Steal focus at startup
On windows, if the user types quickly enough when launching Jolly,
the will steal focus back to whatever application was originally
highlighted.

This means that the Jolly window will appear, but it will not be
focused, and can only be closed by clicking on the window, at which
point the normal Jolly exit criteria allow it to exit.