Added preliminary icon support
Icon support is only implemented  for macos at this point

Icon rendering must use platform-specific APIs which means it must be
duplicated for every supported OS.

Icon lookup is delegated to a separate thread in order to prevent
the system functions from blocking the runtime.

This is because if a Jolly entry location is on a network drive, the
icon lookup APIs can hang for a long period of time.