Bumped iced to 0.10.0
Fixed a bunch of references to iced_native that apparently should not be used.

Also, we seem to be received a spurious window::event::Unfocused
message at startup which was triggering jolly to exit. This is fixed
by ignoring the Unfocused message until we receive at least one
Focused message.