Release GH Action now included example jolly.toml
This means that official builds of the Jolly crate will be distributed
with an example file right next to the executable. Before, if a user
tried to just launch Jolly from the release tarball, they would get an
error from the missing example file and have to copy it over. Now the
example file will already be in the right place.