Fixed flaky windows tests for icons
For some reason, when running icon tests on windows, the default
protocol handler for http and https protocols can fail to load its
icons. This default handler is of course, ms Edge, and I have no idea
why it is flaky.

So as a workaround, the testcase for icon loading on windows now uses
protocol "im", which of course launches MS Teams...